# docs/openapi/swagger.yaml
openapi: 3.1.0
info:
  title: FGD (Federation Governance Dashboard) API
  version: "2.1.0"
  description: |
    REST interface for FGD, the Minecraft NPC swarm management system.
    Routes are mounted as follows:
      - Unversioned surface: `/api/*` — currently served by the v1 route
        tree. `/api/mineflayer` is a special case that always resolves to
        the policy-gated v2 Mineflayer surface (or 503 if the policy
        service failed to initialize).
      - `/api/v1/*` — legacy direct-control surface. Mineflayer endpoints
        under `/api/v1/mineflayer` are DEPRECATED (sunset 2026-07-01).
      - `/api/v2/*` — current surface, including the policy-gated
        Mineflayer routes at `/api/v2/mineflayer`.

    Authentication: protected routes accept either a Bearer JWT
    (`Authorization: Bearer <token>`) or an API key (`X-API-Key`).
servers:
  - url: http://localhost:3000
    description: Local development
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
  schemas:
...
    ErrorEnvelope:
      type: object
      properties:
        success:
          type: boolean
          const: false
        error:
          type: string
        message:
          type: string
      required: [success, error]
    Position:
      type: object
      properties:
        x: { type: number }
        y: { type: number }
        z: { type: number }
      required: [x, y, z]
...
security:
  - BearerAuth: []
  - ApiKeyAuth: []
paths:
...
  /api/auth/login:
    post:
      tags: [Auth]
      summary: Exchange credentials for an access token
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                username: { type: string }
                password: { type: string }
              required: [username, password]
      responses:
        '200':
          description: Access + refresh tokens issued
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean }
                  accessToken: { type: string }
                  refreshToken: { type: string }
        '401':
          description: Invalid credentials
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorEnvelope' }
  /api/auth/me:
    get:
      tags: [Auth]
      summary: Get the currently authenticated user
      responses:
        '200':
          description: Current user
        '401':
          description: Not authenticated
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorEnvelope' }
  /api/auth/refresh:
    post:
      tags: [Auth]
      summary: Exchange a refresh token for a new access token
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                refreshToken: { type: string }
              required: [refreshToken]
      responses:
        '200': { description: New access token issued }
        '401': { description: Invalid or expired refresh token }
...
  /api/npcs:
    get:
      tags: [NPCs]
      summary: List NPCs with pagination
      parameters:
        - in: query
          name: status
          schema: { type: string }
        - in: query
          name: limit
          schema: { type: integer, default: 100 }
        - in: query
          name: offset
          schema: { type: integer, default: 0 }
      responses:
        '200': { description: Paged NPC list }
    post:
      tags: [NPCs]
      summary: Create an NPC
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                id: { type: string }
                role: { type: string }
                npcType: { type: string }
                appearance: { type: object }
                personality: { type: object }
                position: { $ref: '#/components/schemas/Position' }
                autoSpawn: { type: boolean, default: false }
      responses:
        '201': { description: Created }
        '400': { description: Missing role or npcType }
...
  /api/action:
    post:
      tags: [Actions]
      summary: Dispatch any Unified Action Framework action
      responses:
        '200': { description: Dispatched }
        '400': { description: Validation error }

  # ----- Commands (natural-language bot commands) ----------------------
...
  /api/mineflayer:
    get:
      tags: [Mineflayer v2]
      summary: List connected Mineflayer bots (v2)
      responses:
        '200': { description: List }
        '503':
          description: Policy service unavailable — v2 routes not loaded
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorEnvelope' }
...